access denied 2

+++ controllers/ProfController.php
@@ -18,70 +18,78 @@ class ProfController extends Controller
      */
     public function actionCourses()
     {
-        $searchModel = new CoursesSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-        $dataProvider->pagination = ['pageSize' => 15];
+        if(Yii::$app->user->identity->role==1) {
+            $searchModel = new CoursesSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+            $dataProvider->pagination = ['pageSize' => 15];
 
-        return $this->render('courses', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('courses', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            return $this->render('/site/access_denied');
+        }
     }
 
     public function  actionCoursesCreate()
     {
-        $model = new Courses();
+        if(Yii::$app->user->identity->role==1) {
+            $model = new Courses();
 
-        $array = Subjects::find()->asArray()->select('name')->orderBy('name')->all();
-        $subjects = ArrayHelper::getColumn($array, 'name');
+            $array = Subjects::find()->asArray()->select('name')->orderBy('name')->all();
+            $subjects = ArrayHelper::getColumn($array, 'name');
 
-        $subjects_ids = Subjects::find()->asArray()->select('ID')->orderBy('ID')->all(); ///
-        $subjects_ids = ArrayHelper::getColumn($subjects_ids, 'ID');
+            $subjects_ids = Subjects::find()->asArray()->select('ID')->orderBy('ID')->all(); ///
+            $subjects_ids = ArrayHelper::getColumn($subjects_ids, 'ID');
 
-        $subjects = array_combine($subjects_ids,$subjects);
+            $subjects = array_combine($subjects_ids,$subjects);
 
-        //$subjects=array('Предмет 1', 'Предмет 2', 'Предмет 3', 'Предмет 4', 'Предмет 5', 'Предмет 6', 'Предмет 7', 'Предмет 8');
+            //$subjects=array('Предмет 1', 'Предмет 2', 'Предмет 3', 'Предмет 4', 'Предмет 5', 'Предмет 6', 'Предмет 7', 'Предмет 8');
 
-        if ($model->load(Yii::$app->request->post()) && $model->validate()) {
-            // данные в $model удачно проверены
+            if ($model->load(Yii::$app->request->post()) && $model->validate()) {
+                // данные в $model удачно проверены
 
-            $coursesName = Html::encode($model->name);
-            $coursesPract = Html::encode($model->pract);
-            $coursesWorklect = Html::encode($model->worklect);
-            $coursesTeorlect = Html::encode($model->teorlect);
-            $coursesSubject ='';
-            $subjectCount = count($model->subject);
-            $i=0;
-            foreach ($model->subject as $subject) {
-                $i++;
-                if($i==$subjectCount) {
-                    $coursesSubject = $coursesSubject . Html::encode($subject);
-                } else {
-                    $coursesSubject = $coursesSubject . Html::encode($subject.", ");
+                $coursesName = Html::encode($model->name);
+                $coursesPract = Html::encode($model->pract);
+                $coursesWorklect = Html::encode($model->worklect);
+                $coursesTeorlect = Html::encode($model->teorlect);
+                $coursesSubject ='';
+                $subjectCount = count($model->subject);
+                $i=0;
+                foreach ($model->subject as $subject) {
+                    $i++;
+                    if($i==$subjectCount) {
+                        $coursesSubject = $coursesSubject . Html::encode($subject);
+                    } else {
+                        $coursesSubject = $coursesSubject . Html::encode($subject.", ");
+                    }
                 }
+
+                $model->name = $coursesName;
+                $model->pract = $coursesPract;
+                $model->worklect = $coursesWorklect;
+                $model->teorlect = $coursesTeorlect;
+                $model->subject = $coursesSubject;
+
+                $model->save();
+
+                //return $this->redirect(['courses-create', 'id' => $model->ID]);
+                return $this->render('courses-create', [
+                    'model' => $model,
+                    'operation' => 'created',
+                    'subjects' => $subjects,
+
+                ]);
+            } else {
+                // либо страница отображается первый раз, либо есть ошибка в данных
+                return $this->render('courses-create', [
+                    'model' => $model,
+                    'subjects' => $subjects,
+                ]);
             }
-
-            $model->name = $coursesName;
-            $model->pract = $coursesPract;
-            $model->worklect = $coursesWorklect;
-            $model->teorlect = $coursesTeorlect;
-            $model->subject = $coursesSubject;
-
-            $model->save();
-
-            //return $this->redirect(['courses-create', 'id' => $model->ID]);
-            return $this->render('courses-create', [
-                'model' => $model,
-                'operation' => 'created',
-                'subjects' => $subjects,
-
-            ]);
         } else {
-            // либо страница отображается первый раз, либо есть ошибка в данных
-            return $this->render('courses-create', [
-                'model' => $model,
-                'subjects' => $subjects,
-            ]);
+            return $this->render('/site/access_denied');
         }
     }
 
@@ -94,10 +102,14 @@ class ProfController extends Controller
      */
     protected function findModel($id)
     {
-        if (($model = Courses::findOne($id)) !== null) {
-            return $model;
+        if(Yii::$app->user->identity->role==1) {
+            if (($model = Courses::findOne($id)) !== null) {
+                return $model;
+            } else {
+                throw new NotFoundHttpException('Сторінку не знайдено.');
+            }
         } else {
-            throw new NotFoundHttpException('Сторінку не знайдено.');
+            return $this->render('/site/access_denied');
         }
     }
 
@@ -109,9 +121,13 @@ class ProfController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->identity->role==1) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['courses']);
+            return $this->redirect(['courses']);
+        } else {
+            return $this->render('/site/access_denied');
+        }
     }
 
     /**
@@ -121,44 +137,48 @@ class ProfController extends Controller
      * @return mixed
      */
     public function actionUpdate($id) {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->identity->role==1) {
+            $model = $this->findModel($id);
 
-        $array = Subjects::find()->asArray()->select('name')->orderBy('name')->all();
-        $subjects = ArrayHelper::getColumn($array, 'name');
+            $array = Subjects::find()->asArray()->select('name')->orderBy('name')->all();
+            $subjects = ArrayHelper::getColumn($array, 'name');
 
-        //$subjects=array('Предмет 1', 'Предмет 2', 'Предмет 3', 'Предмет 4', 'Предмет 5', 'Предмет 6', 'Предмет 7', 'Предмет 8');
+            //$subjects=array('Предмет 1', 'Предмет 2', 'Предмет 3', 'Предмет 4', 'Предмет 5', 'Предмет 6', 'Предмет 7', 'Предмет 8');
 
-        if ($model->load(Yii::$app->request->post()) && $model->validate()) {
+            if ($model->load(Yii::$app->request->post()) && $model->validate()) {
 
-            $coursesName = Html::encode($model->name);
-            $coursesPract = Html::encode($model->pract);
-            $coursesWorklect = Html::encode($model->worklect);
-            $coursesTeorlect = Html::encode($model->teorlect);
-            $coursesSubject='';
-            foreach ($model->subject as $subject) {
-                $coursesSubject = $coursesSubject . Html::encode($subject.", ");
+                $coursesName = Html::encode($model->name);
+                $coursesPract = Html::encode($model->pract);
+                $coursesWorklect = Html::encode($model->worklect);
+                $coursesTeorlect = Html::encode($model->teorlect);
+                $coursesSubject='';
+                foreach ($model->subject as $subject) {
+                    $coursesSubject = $coursesSubject . Html::encode($subject.", ");
+                }
+
+                $model->name = $coursesName;
+                $model->pract = $coursesPract;
+                $model->worklect = $coursesWorklect;
+                $model->teorlect = $coursesTeorlect;
+                $model->subject = $coursesSubject;
+
+                $model->update();
+
+                //return $this->redirect(['update', 'id' => $model->ID, 'operation' => 'updated']);
+                return $this->render('update', [
+                    'model' => $model,
+                    'operation' => 'updated',
+                    'id' => $model->ID,
+                    'subjects' => $subjects,
+                ]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                    'subjects' => $subjects,
+                ]);
             }
-
-            $model->name = $coursesName;
-            $model->pract = $coursesPract;
-            $model->worklect = $coursesWorklect;
-            $model->teorlect = $coursesTeorlect;
-            $model->subject = $coursesSubject;
-
-            $model->update();
-
-            //return $this->redirect(['update', 'id' => $model->ID, 'operation' => 'updated']);
-            return $this->render('update', [
-                'model' => $model,
-                'operation' => 'updated',
-                'id' => $model->ID,
-                'subjects' => $subjects,
-            ]);
         } else {
-            return $this->render('update', [
-                'model' => $model,
-                'subjects' => $subjects,
-            ]);
+            return $this->render('/site/access_denied');
         }
     }
